Aggregate or scale one subtree's children values to the parent level

Aggregate or scale one subtree's children values to the parent level

agg_subtree(
  dt,
  by_id_cols,
  value_cols,
  col_stem,
  col_type,
  agg_function,
  subtree
)

scale_subtree(
  dt,
  by_id_cols,
  value_cols,
  col_stem,
  col_type,
  agg_function,
  subtree
)

Arguments

dt
[data.table()]
Data to be aggregated or scaled.

by_id_cols
[character()]
The id_cols without the agg/scale variable(s) included.

value_cols
[character()]
Value columns that should be aggregated.

col_stem
[character(1)]
The name of the variable to be aggregated or scaled over. If aggregating an 'interval' variable should not include the '_start' or '_end' suffix.

col_type
[character(1)]
The type of variable that is being aggregated or scaled over. Can be either 'categorical' or 'interval'.

agg_function
[function()]
Function to use when aggregating, can be either sum (for counts) or prod (for probabilities).

subtree
one subtree of the list returned by create_agg_subtrees() or create_scale_subtrees().

Value

[data.table()] with same input columns for requested aggregates or with scaled values for the given subtree. scale_subtree() changes the name of the value_cols to {value_cols}_scaled.